requirements.txt
flask==3.1.2
google-generativeai==0.1.0
python-dotenv==1.0.0
gunicorn==21.2.0